[WHD-276] Fix: Migration image gallery saver to image gallery saver plus
Image gallery saver의 버전으로 인해 image gallery saver plus로 변경
- General 함수 분기화
