+++ .github/workflows/flutter_apk_build.yml
@@ -2,9 +2,14 @@ name: Flutter APK Build
 
 on:
   push:
-    branches: [ "main" ]
+    branches:
+      - "main"
+      - "develop"
   pull_request:
-    branches: [ "main" ]
+    branches:
+      - "main"
+      - "develop"
+  workflow_dispatch:
 
 jobs:
   build:
