+++ lib/view/club_info/club_info_promotion_page.dart
@@ -2,12 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 import 'package:material_symbols_icons/symbols.dart';
-// ignore: implementation_imports
-import 'package:widgets_to_png/src/entity/image_converter.dart';
-import 'package:widgets_to_png/widgets_to_png.dart';
 import 'package:woohakdong/view/themes/theme_context.dart';
 import 'package:woohakdong/view_model/group/group_provider.dart';
 
+import '../../service/general/general_image.dart';
 import '../../service/general/general_functions.dart';
 import '../club_register/components/club_register_qr_card.dart';
 import '../club_register/components/club_register_url_card.dart';
@@ -28,7 +26,7 @@ class ClubInfoPromotionPage extends ConsumerWidget {
         actions: [
           IconButton(
             onPressed: () async {
-              await ImageConverter.saveWidgetToGallery(
+              await GeneralImage.convertWidgetToPng(
                 key: _widgetToPngKey,
                 fileName: '우학동 QR 카드 ${DateTime.now().millisecondsSinceEpoch}.png',
               );
@@ -57,8 +55,8 @@ class ClubInfoPromotionPage extends ConsumerWidget {
                   border: Border.all(color: context.colorScheme.surfaceContainer),
                 ),
                 padding: const EdgeInsets.all(defaultPaddingS / 2),
-                child: WidgetToPng(
-                  keyToCapture: _widgetToPngKey,
+                child: RepaintBoundary(
+                  key: _widgetToPngKey,
                   child: ClubRegisterQrCard(groupInfo: groupInfo),
                 ),
               ),
